Diffusing t1 instead of t1a
Don't recall exactly what I was trying to accomplish here.

t1a is an unadvected version of t1. I was sending that to diffusion, but then never using the results, so diffusion wasn't occurring. The cobwebs of my mind say that this might have been part of an earlier attempt at parallelization. I need to search through my notes to verify.

That, or, I am supposed to diffuse the nonadvected temperatures, but I just forget to combine those results with the advection for pgf_bouyancy.